Created basic team page with header
